Create separate header components, change headline
